refactor: 적절한 패키지로 클래스 이동
정책상 사라진 model 패키지 내 클래스들 이동
