+++ src/main/java/com/zerobase/everycampingbackend/domain/auth/issuer/JwtIssuer.java
@@ -1,8 +1,8 @@
 package com.zerobase.everycampingbackend.domain.auth.issuer;
 
 
-import com.zerobase.everycampingbackend.domain.auth.model.JwtDto;
-import com.zerobase.everycampingbackend.domain.auth.model.UserVo;
+import com.zerobase.everycampingbackend.domain.auth.dto.JwtDto;
+import com.zerobase.everycampingbackend.domain.auth.dto.UserVo;
 import com.zerobase.everycampingbackend.domain.auth.util.Aes256Util;
 import com.zerobase.everycampingbackend.exception.CustomException;
 import com.zerobase.everycampingbackend.exception.ErrorCode;
