--- src/main/java/com/zerobase/everycampingbackend/domain/auth/model/UserType.java
@@ -1,5 +1,0 @@
-package com.zerobase.everycampingbackend.domain.auth.model;
-
-public enum UserType {
-    CUSTOMER, SELLER, ADMIN
-}
