+++ src/main/java/com/zerobase/everycampingbackend/domain/auth/type/UserType.java
@@ -1,0 +1,5 @@
+package com.zerobase.everycampingbackend.domain.auth.type;
+
+public enum UserType {
+    CUSTOMER, SELLER, ADMIN
+}
